feat: Update login form to use NavigationDataManager for screen navigation
Refactor the listener in the login form to use the NavigationDataManager to retrieve the previous screen route and arguments. This allows for more flexible navigation after a successful login, as the user will be redirected to the previous screen instead of a fixed route. The previous screen route and arguments are stored in the NavigationDataManager when navigating to the login screen. This change improves the user experience and makes the login flow more intuitive.